+++ entrytracker.h
@@ -88,9 +88,9 @@ public:
                     in_desc));
     }
 
-    int get_field_id(const std::string& in_name);
-    std::string get_field_name(int in_id);
-    std::string get_field_description(int in_id);
+    uint16_t get_field_id(const std::string& in_name);
+    std::string get_field_name(uint16_t in_id);
+    std::string get_field_description(uint16_t in_id);
 
     // Generate a shared field instance, using the builder
     template<class T> std::shared_ptr<T> get_shared_instance_as(const std::string& in_name) {
@@ -98,10 +98,10 @@ public:
     }
     std::shared_ptr<tracker_element> get_shared_instance(const std::string& in_name);
 
-    template<class T> std::shared_ptr<T> get_shared_instance_as(int in_id) {
+    template<class T> std::shared_ptr<T> get_shared_instance_as(uint16_t in_id) {
         return std::static_pointer_cast<T>(get_shared_instance(in_id));
     }
-    std::shared_ptr<tracker_element> get_shared_instance(int in_id);
+    std::shared_ptr<tracker_element> get_shared_instance(uint16_t in_id);
 
     // Register a serializer for auto-serialization based on type
     void register_serializer(const std::string& type, std::shared_ptr<tracker_element_serializer> in_ser);
@@ -116,9 +116,9 @@ public:
 
     // Optional per-field-id transforms for search functions, must use the search workers or be called
     // manually
-    void register_search_xform(int in_field_id, std::function<void (std::shared_ptr<tracker_element>,
+    void register_search_xform(uint16_t in_field_id, std::function<void (std::shared_ptr<tracker_element>,
                 std::string& mapped_str)> in_xform);
-    void remove_search_xform(int in_field_id);
+    void remove_search_xform(uint16_t in_field_id);
     // Apply a search transform to a field, returning 'true' if the field was transformable, 
     // and placing the results in mapped_str
     bool search_xform(std::shared_ptr<tracker_element> elem, std::string& mapped_str);
@@ -131,7 +131,7 @@ protected:
 
     struct reserved_field {
         // ID we assigned
-        int field_id;
+        uint16_t field_id;
 
         // Readable metadata
         std::string field_name;
@@ -142,11 +142,11 @@ protected:
     };
 
     robin_hood::unordered_node_map<std::string, std::shared_ptr<reserved_field> > field_name_map;
-    robin_hood::unordered_node_map<int, std::shared_ptr<reserved_field> > field_id_map;
+    robin_hood::unordered_node_map<uint16_t, std::shared_ptr<reserved_field> > field_id_map;
     robin_hood::unordered_node_map<std::string, std::shared_ptr<tracker_element_serializer> > serializer_map;
 
     // Field IDs to optional search xform function
-    robin_hood::unordered_node_map<int, std::function<void (std::shared_ptr<tracker_element>, 
+    robin_hood::unordered_node_map<uint16_t, std::function<void (std::shared_ptr<tracker_element>, 
             std::string& mapped_str)>> search_xform_map;
 
     void tracked_fields_endp_handler(std::shared_ptr<kis_net_beast_httpd_connection> con);
